requirements.txt
asyncio  
speechrecognition  
g4f[all]
python-telegram-bot  
pydub  
edge-tts  